// libs/platform/list/access-data/src/lib/enums/statusses/announcement-status.enum.ts
export enum AnnouncementStatusIdEnum {
    Active = 1,
    Draft = 2,
    Inactive = 3
}

export enum AnnouncementStatusNameEnum {
    Active = 'active',
    Draft = 'draft',
    Inactive = 'inactive'
}

export enum AnnouncementStatusTranslateEnum {
    Active = 'statuses.active',
    Draft = 'statuses.draft',
    Inactive = 'statuses.inactive'
}